requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20231117
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
scipy>=1.11.0
openai>=1.3.0
google-cloud-speech 